Refactored robotpy_toolbox.ts: Moved everything to python_data_toolbox.ts, but instead of always using robotPyData, getToolboxCategoriees takes a pythonData argument. Changed robotpy_toolbox getToolboxCategories to call python_data_toolbox getToolboxCategories with robotPyData
