+++ frontend/blocks/mrc_call_python_function.ts
@@ -44,7 +44,7 @@ import { MRC_STYLE_FUNCTIONS } from '../themes/styles'
 import * as toolboxItems from '../toolbox/items';
 import * as storageModule from '../storage/module';
 import * as storageModuleContent from '../storage/module_content';
-import { makeOneContents } from '../toolbox/robotpy_toolbox';
+import { makeOneContents } from '../toolbox/python_data_toolbox';
 
 
 // A block to call a python function.
